cleans app component
