+++ Bookcatalog.WebReact/src/types/authInfo.ts
@@ -4,16 +4,16 @@ export type authInfo = {
     token?: string,
     refreshToken?: string,
     isError?: boolean
-}
+};
 
 export type UserInfo = {
     userName?: string,
     firstName?: string,
     lastName?: string,
     phoneNum?: string
-}
+};
 
 export type Login = {
     username: string,
     password: string
-}
+};
