+++ Bookcatalog.WebReact/src/utils/auth.ts
@@ -1,11 +1,6 @@
 import axios from 'axios';
 import { getApiUrl, refreshTokenUrl } from 'config/url';
 
-/* type RefreshTokenState = {
-    isRefreshed: boolean,
-    isComplete: boolean
-} */
-
 export const setAuthTokenHeader = (token: string) => {
     if (token) {
         axios.defaults.headers.common["Authorization"] = `Bearer ${token}`;
@@ -13,53 +8,6 @@ export const setAuthTokenHeader = (token: string) => {
     else
         delete axios.defaults.headers.common["Authorization"];
 }
-
-axios.interceptors.request.use((config) => {
-    const token = localStorage.getItem("bookCatalogToken");
-
-    if (token) {
-        config.headers.Authorization = `Bearer ${token}`;
-    }
-    else {
-        config.headers.Authorization = null;
-    }
-
-    return config;
-}, (error) => {
-    return Promise.reject(error);
-});
-
-axios.interceptors.response.use(response => {
-    localStorage.removeItem("refreshingTokenRequested");
-
-    return response;
-}, async error => {
-
-
-    if (error.response.status === 401) {
-
-        if (!localStorage.getItem("refreshingTokenRequested")) {
-
-            localStorage.setItem("refreshingTokenRequested", 'true');
-
-            console.log('axios.interceptor=>response.status === 401');
-            const tokenRefreshSuccessful = await refreshToken();
-
-            if (!tokenRefreshSuccessful) {
-
-                //TODO: move to function
-                localStorage.removeItem("bookCatalogToken");
-                localStorage.removeItem("bookCatalogRefreshToken");
-                localStorage.removeItem("bookCatalogUserName");
-                localStorage.removeItem("refreshingTokenRequested");
-
-                window.location.hash = '/login'
-            }
-        }
-    }
-
-    return error;
-});
 
 export const refreshToken = async () => {
 
@@ -91,10 +39,4 @@ export const refreshToken = async () => {
 
 export const isUserAuthenicated = (): boolean => {
     return localStorage.getItem("bookCatalogToken") !== '' && localStorage.getItem("bookCatalogToken") !== null;
-    /* if(localStorage.getItem("bookCatalogToken")){
-        return true;
-    }
-    else{
-        return false;
-    } */
 }
